# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GRT Transit API",
    "description": "A proxy service that ingests GTFS-Realtime Protocol Buffer feeds from Grand River Transit (GRT) and exposes them as RESTful JSON endpoints.\n\nThis API is designed in accordance with the [Microsoft Graph REST API Guidelines](https://github.com/microsoft/api-guidelines/blob/vNext/graph/GuidelinesGraph.md).",
    "contact": {
      "name": "John Steel",
      "email": "john@steelcomputers.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local/Relative Server"
    }
  ],
  "paths": {
    "/routes/{route_id}/vehicles": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get Vehicles By Route",
        "description": "Returns a collection of currently active vehicles specifically assigned to the requested route ID.",
        "operationId": "get_vehicles_by_route_routes__route_id__vehicles_get",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get All Alerts",
        "description": "Returns a full collection of all currently active transit alerts across the entire GRT network.",
        "operationId": "get_all_alerts_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/routes/{route_id}/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get Alerts By Route",
        "description": "Returns a collection of active transit alerts filtered dynamically to only ones impacting the requested route ID.",
        "operationId": "get_alerts_by_route_routes__route_id__alerts_get",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routes/{route_id}/trips": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get Trips By Route",
        "description": "Returns an array of ETA updates and stop schedules for all presently active vehicle trips on the requested route ID.",
        "operationId": "get_trips_by_route_routes__route_id__trips_get",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "Get All Stations",
        "description": "Returns a collection of stops and stations on the GRT network. Defaults to `locationType=0` (individual stops/platforms). Pass `locationType=1` to retrieve parent station records instead. Supports `$top` and `$skip` for pagination.",
        "operationId": "get_all_stations_stations_get",
        "parameters": [
          {
            "name": "$top",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "$Top"
            }
          },
          {
            "name": "$skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "$Skip"
            }
          },
          {
            "name": "locationType",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Locationtype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stations/{station_id}": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "Get Station By Id",
        "description": "Look up a specific station or stop by its unique ID.",
        "operationId": "get_station_by_id_stations__station_id__get",
        "parameters": [
          {
            "name": "station_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "Vehicles",
      "description": "Endpoints regarding real-time vehicle positioning."
    },
    {
      "name": "Alerts",
      "description": "Endpoints regarding active GTFS-RT service alerts."
    },
    {
      "name": "Trips",
      "description": "Endpoints for forecasting vehicle arrivals at stops and schedule deviations."
    },
    {
      "name": "Stations",
      "description": "Endpoints for static definitions of transit hubs, platforms, and specific stops."
    }
  ]
}